[#19] Refactor : 코드 리팩토링 및 버그 Fix
* [#19] Refactor : 메인페이지 컴포넌트 분리 및 목 데이터 연결

- url 로 사용중인 더미데이터 json 파일 추가 및 변경
- 메인페이지 컴포넌트 분리

* [#19] Refactor : 불필요한 컴포넌트 삭제 및 ChakraProvider 파일 위치 이동

* [#19] Refactor : toast추가 상품삭제 confirm

* [#19] Refactor : ProductModal 컴포넌트 분리

* [#19] Refactor : Slider / CheckBox 분리 후 React.memo 적용

* [#19] Refactor : Custom Toast 구현

---------

+++ src/components/reservation/Reservation.tsx
@@ -14,6 +14,8 @@ function Reservation({
 	setReservations,
 }: IReservationProps) {
 	const handleDelete = (productIdx: number) => {
+		if (!window.confirm('삭제 하시겠습니까?')) return;
+
 		let reservations = JSON.parse(localStorage.getItem('reservations') as string);
 		reservations = reservations.filter((reservation: IReservation) => reservation.idx !== productIdx);
 		localStorage.setItem('reservations', JSON.stringify(reservations));
